build.sh reverted back to using scripts/adjust-pkgconfig.sh because pkgconf still isn't doing what we need when PKG_CONFIG_SYSROOT_DIR is used
cfg/build/package/argp-standalone
the build script was not installing the generated files.

cfg/build/package/*
the adjust-pkgconfig.sh script now requires a new parameter in
order to make the adjusted files use our fake sysroot dir. We do
that to avoid clutting the pkgconf search paths with a new dir for
each package we build. And, because of that, we had to change
pretty much every package that calls adjust-pkgconfig.sh

cfg/script/
	adjust-pkgconfig.sh
	build-xorg-lib.sh
	build-xorg-proto.sh
		see above.